projects/auto_ip/requirements.txt
selenium==3.141.0
python-dotenv=0.12.0
requests=2.23.0
